+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - `@use-gpu/workbench`: Improved WebGPU resource labels (for error messages and inspector)
 - `@use-gpu/workbench`: Optional `depthHistory` for `<RenderTarget>`
 - `@use-gpu/workbench`: Add `decimate` and `converge` options to `<Loop>` to skip or converge frames.
+- `@use-gpu/workbench`: Fix `<Environment>` `gain` not working in some cases
 
 0.13.1
 - `@use-gpu/app`: Add path tracing + debug picking example.